golang/xenlight: re-name Bitmap marshaling functions
Re-name and modify signature of toGo function to fromC. The reason for
using 'fromC' rather than 'toGo' is that it is not a good idea to define
methods on the C types. Also, add error return type to Bitmap's toC function.

Finally, as code-cleanup, re-organize the Bitmap type's comments as per
Go conventions.

Signed-off-by: Nick Rosbrook <rosbrookn@ainfosec.com>
Acked-by: George Dunlap <george.dunlap@citrix.com>
Message-Id: <4c635331e4a5171e99388391b536d5961ea2fd96.1573840474.git.rosbrookn@ainfosec.com>