xen/arm: remove physical timer offset
The physical timer traps apply an offset so that time starts at 0 for
the guest. However, this offset is not currently applied to the physical
counter. Per the ARMv8 Arch Reference Manual, the offset between the
physical timer and counter should be 0. This removes the offset to make
the timer and counter consistent.

Xen time is at offset boot_count from the physical counter, so we need
to take this into account when reading/writing to CNTP_CVAL.

Signed-off-by: Jeff Kubascik <jeff.kubascik@dornerworks.com>
Message-Id: <20191125161419.75909-1-jeff.kubascik@dornerworks.com>