xen/arm: initialize vpl011 flag register
The tx/rx fifo flags were not set when the vpl011 is initialized. This
is a problem for certain guests that are operating in polled mode, as a
guest will generally check the rx fifo empty flag to determine if there
is data before doing a read. The result is a continuous spam of the
message "vpl011: Unexpected IN ring buffer empty" before the first valid
character is received. This initializes the flag status register to the
default specified in the PL011 technical reference manual.

Signed-off-by: Jeff Kubascik <jeff.kubascik@dornerworks.com>

Changes in v2:
- Moved uartfr initialization to later point in function after potential
return/failure points
Message-Id: <20191125205800.64602-1-jeff.kubascik@dornerworks.com>
Acked-by: Julien Grall <julien@xen.org>