x86: store cr4 during suspend/resume
Currently cr4 is not cached before suspension, and mmu_cr4_features is
used in order to restore the expected cr4 value. This is correct so
far because the tasklet that executes the suspend/resume code is
running in the idle vCPU context.

In order to make the code less fragile, explicitly save the current
cr4 value before suspension, so that it can be restored afterwards.
This ensures that the cr4 value cached in the cpu_info doesn't get out
of sync after resume from suspension.

Suggested-by: Jan Beulich <jbeulich@suse.com>
Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>
Message-Id: <20191209180638.47305-1-roger.pau@citrix.com>
Reviewed-by: Jan Beulich <jbeulich@suse.com>