xen/tools: VM forking toolstack side
Add necessary bits to implement "xl fork-vm", "xl fork-launch-dm" and
"xl fork-reset" commands. The process is split in two to allow tools needing
access to the new VM as fast as possible after it was forked. It is expected
that under certain use-cases the second command that launches QEMU will be
skipped entirely.

Signed-off-by: Tamas K Lengyel <tamas.lengyel@intel.com>
Message-Id: <122128bd25e184d310a0a8af0f3c1e1c272d2fe1.1576697796.git.tamas.lengyel@intel.com>