automation: Cache sub-project git tree in build jobs
GitLab have a caching capability, see [1]. Let's use it to avoid using
Internet too often.

The cache is setup so that when xen.git/Config.mk is changed, the
cache will need to be recreated. This has been chosen because that is
where the information about how to clone sub-project trees is encoded
(revisions). That may not work for qemu-xen tree which usually is
`master', but that should be fine for now.

The cache is populated of "git bundle" which will contain a mirror of
the original repo, and can be cloned from. If the bundle exist, the
script have the Xen makefiles clone from it, otherwise it will clone
from the original URL and the bundles will be created just after.

We have more than one runner in GitLab, and no shared cache between
them, so every build jobs will be responsible to create the cache.

[1] https://docs.gitlab.com/ee/ci/yaml/README.html#cache

Signed-off-by: Anthony PERARD <anthony.perard@citrix.com>
Message-Id: <20191219144217.305851-3-anthony.perard@citrix.com>