x86/boot: Clean up the trampoline transition into Long mode
The jmp after setting %cr0 is redundant with the following ljmp.

The CPUID to protect the jump to higher mappings was inserted due to an
abundance of caution/paranoia before Spectre was public.  There is not a
matching protection in the S3 resume path, and there is nothing
interesting in memory at this point.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
Message-Id: <20200102145953.6503-1-andrew.cooper3@citrix.com>
Acked-by: Jan Beulich <jbeulich@suse.com>